bot/requirements.txt
discord.py
wavelink
python-dotenv
aiohttp